% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darts1.R
\name{inside}
\alias{inside}
\title{Test whether each dart landed within the dartboard}
\usage{
inside(darts)
}
\arguments{
\item{darts}{nx2 matrix containing dart locations}
}
\value{
Boolean vector of length n stating whether each dart landed in dartboard.
}
\description{
Did the darts land within the dartboard?
}
\details{
Return TRUE/FALSE for each element of the darts
}
\author{
Stephen Eglen
}
