% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darts1.R
\name{plot_dartboard}
\alias{plot_dartboard}
\title{Plot the dartboard and darts}
\usage{
plot_dartboard(darts, inside)
}
\arguments{
\item{darts}{nx2 matrix containing dart locations}

\item{inside}{Vector of length n stating whether each dart landed in dartboard.}
}
\value{
Nothing
}
\description{
Plot the dartboard with the darts showing which are inside/outside.
}
\author{
Stephen Eglen
}
