% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darts1.R
\name{throw_n_darts}
\alias{throw_n_darts}
\title{Simulate throwing darts at a board.}
\usage{
throw_n_darts(n)
}
\arguments{
\item{n}{The number of darts to thrown}
}
\value{
A nx2 matrix with each row containing (x,y) location of dart.
}
\description{
Simulate throwing N darts at a dartboard
}
\details{
The x and y locations are drawn independently from a uniform
distribution -1 to +1.
}
\author{
Stephen Eglen
}
